modularized base64 image processing to its own module and updated gitignore
