+++ RideRecon/.gitignore
@@ -3,6 +3,9 @@
 # dependencies
 node_modules/
 
+# Python
+__pycache__/
+
 # Expo
 .expo/
 dist/
