+++ RideRecon/services/experts/gemini.py
@@ -1,8 +1,7 @@
 import vertexai
 from vertexai.generative_models import GenerativeModel, SafetySetting, Part
-import vertexai.preview.generative_models as models_preview
+from base64_funcs import *
 import os
-import base64
 from pathlib import Path
 from dotenv import load_dotenv
 from json import loads
@@ -29,40 +28,6 @@ def resize_image(image_path, max_size=(1024, 1024), quality=85):
         print(f"Error resizing image: {e}")
         with open(image_path, "rb") as f:
             return f.read()  # Return original as fallback
-
-def image_to_base64(image_path):
-    """
-    Convert an image to base64 encoded string
-    
-    Args:
-        image_path (str): Path to the image file
-        
-    Returns:
-        str: Base64 encoded string
-    """
-    try:
-        # Handle both relative and absolute paths
-        absolute_path = os.path.abspath(image_path)
-        
-        # Check if file exists
-        if not os.path.exists(absolute_path):
-            # Try to resolve path relative to project root
-            project_root = Path(__file__).parent.parent.parent  # Go up to RideRecon folder
-            alternate_path = project_root / "assets" / "images" / os.path.basename(image_path)
-            
-            if os.path.exists(alternate_path):
-                absolute_path = str(alternate_path)
-            else:
-                raise FileNotFoundError(f"Image not found at {image_path} or {alternate_path}")
-        
-        print(f"Reading image from {absolute_path}")
-        with open(absolute_path, "rb") as image_file:
-            binary_data = image_file.read()
-            base64_encoded = base64.b64encode(binary_data)
-            return base64_encoded.decode('utf-8')
-    except Exception as e:
-        print(f"Error converting image to base64: {e}")
-        raise
 
 def load_environment():
     script_dir = Path(os.path.dirname(os.path.abspath(__file__)))
